+++ static/js/preload.js
@@ -138,6 +138,7 @@ contextBridge.exposeInMainWorld('electronAPI', {
   registerVrmHideShortcut: (key) => ipcRenderer.invoke('register-vrm-hide-shortcut', key),
   unregisterVrmHideShortcut: () => ipcRenderer.invoke('unregister-vrm-hide-shortcut'),
   vrmWander: (opts) => ipcRenderer.invoke('vrm-wander', opts),
+  summonFriend: (opts) => ipcRenderer.invoke('summon-vrm-friend', opts),
   readDirectory: (dirPath) => ipcRenderer.invoke('read-directory', dirPath),
   deleteWorkspaceFile: (filePath) => ipcRenderer.invoke('delete-workspace-file', filePath),
   uploadToWorkspace: (targetDirPath, sourceFilePaths) => ipcRenderer.invoke('upload-to-workspace', { targetDirPath, sourceFilePaths }),